Remove dialog_num from window structure again
This was introduced by commit 796a5494c751fdbf80e845729f55573e37bff98b.
It makes the window structure incompatible. Since that is exposed to modules,
and the 106gpl archive contains some that currently cannot be recompiled,
we must not do that.

bubblegm.c now contains an array with the predefined window numbers.